вынес в отдельные компоненты compound components
