+++ posts-app/src/shared/context/ModalContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from "react"
+
+const ModalContext = createContext<{ toogle: () => void } | undefined>(undefined)
+
+export default ModalContext
